[refactor] deleted getPosts and getUsers which were previouslt replaced by redux
